Refactor xPDOObject->remove() to use key=>value for primary key criteria
When using string primary keys that are stored as binary data in the database, xPDO will falsely detect the value as either a SQL operator or a SQL injection attempt when encountering certain valid characters. This requires that these types of primary keys be represented in the key => value criteria format instead of just the primary key value. The remove() method was not using this format and causing issues when used with these kinds of string primary keys.