Improve PSR-4 support by allowing namespace prefixes to be removed from package name (#168)
Previously, in order to use PSR-4, you could not include the entire namespace in the package name when using `setPackage()` or `addPackage()` with a namespace prefix. This change will allow you to specify the entire namespace within the package name and pass the namespace prefix into the method calls so they know not to append the entire namespace to the path when searching for the metadata file.