Fix xPDOFileCache->delete() when a directory with same cache key exists
Other xPDOCache implementations check for the multiple_object_delete option before attempting to recursively remove cache entries that match the cache key provided to the delete() method. xPDOFileCache was not doing this, and further, if a directory existed with the cache key to be deleted, the directory was emptied of cache entries but the actual cache entry that was meant to be deleted was left in place.

This commit also adds a constant entry for the option, switches to using getOption() for the checks, and makes the logic consistent across all implementations.

Resolves #170 for 3.x branch